+++ docs/CHANGELOG.md
@@ -43,7 +43,7 @@ You can find example Subsonic responses from an official server and Ampache serv
 * webplayer
   * Add a button next to the playlist to allow looping after the last song
   * If you enable playlist loop do not remove previous tracks
-* Database 600034
+* Database 600036
   * Add preference `webplayer_removeplayed`, Remove tracks before the current playlist item in the webplayer when played
   * Drop channel table
   * Add `total_skip` to podcast table
@@ -74,6 +74,8 @@ You can find example Subsonic responses from an official server and Ampache serv
   * Extend `object_type` enum list on `rating` table
   * Convert `object_type` to an enum on `user_flag` table
   * Convert `object_type` to an enum on `image` table
+  * Add `enabled` to podcast_episode table
+  * Update user `play_size` and catalog `size` fields to megabytes (Stop large catalogs overflowing 32bit ints)
 * Config version 67
   * Drop Channels from config
   * Reset the art_order defaults (replace lastfm with spotify)